chore: renamed legacy folder and paths
