+++ frontend/src/components/CreateNewIndexer/CreateNewIndexer.js
@@ -1,4 +1,4 @@
-import Editor from '@/components/Editor/Legacy/Editor';
+import Editor from '@/components/Editor/EditorComponents/Editor';
 
 const CreateNewIndexer = () => {
   return <Editor actionButtonText="Create New Indexer" theme="vs-dark" />;
