+++ frontend/src/components/Editor/EditorComponents/Editor.tsx
