+++ frontend/src/components/Editor/EditorComponents/ResizableLayoutEditor.jsx
