+++ frontend/src/components/Editor/EditorComponents/block_details.js
