+++ frontend/src/components/Editor/EditorComponents/custom.d.ts
